+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/formatters/base_formatter.rb
@@ -3,6 +3,9 @@ module RubyLsp
     class BaseFormatter
       def self.call(source, uri, message_queue, index)
         new(message_queue, index).call(source, uri)
+      rescue => e
+        $stderr.puts("error in #{self.class}: #{e.message} : #{e.backtrace.first}")
+        source
       end
 
       def initialize(message_queue, index)