+++ ruby-lsp-ree/lib/ruby_lsp/ruby_lsp_ree/formatters/unused_links_formatter.rb
@@ -7,38 +7,34 @@ module RubyLsp
     class UnusedLinksFormatter < BaseFormatter
       include RubyLsp::Ree::ReeLspUtils
 
+      attr_reader :editor, :dsl_parser
+
       def call(source, _uri)
         parsed_doc = RubyLsp::Ree::ParsedDocumentBuilder.build_from_source(source)
         return source if !parsed_doc
 
         parsed_doc.parse_links
 
-        editor = RubyLsp::Ree::ReeSourceEditor.new(source)
-        dsl_parser = RubyLsp::Ree::ReeDslParser.new(parsed_doc, @index)
+        @editor = RubyLsp::Ree::ReeSourceEditor.new(source)
+        @dsl_parser = RubyLsp::Ree::ReeDslParser.new(parsed_doc, @index)
 
         links_count = parsed_doc.link_nodes.size
 
         removed_links = 0
 
         parsed_doc.link_nodes.each do |link_node|
-          removed_imports = 0
+          remove_imports = []
 
           if link_node.has_import_section?
-            link_node.imports.each do |link_import|
-              # TODO extract condition
-              next if editor.contains_link_import_usage?(link_node, link_import) || dsl_parser.contains_object_usage?(link_import)
-              
-              editor.remove_link_import(link_node, link_import)
-              removed_imports += 1
-            end
+            remove_imports = link_node.imports.reject{ |imp| import_is_used?(link_node, imp) }
+            editor.remove_link_imports(link_node, remove_imports)
 
-            if link_node.imports.size == removed_imports
+            if link_node.imports.size == remove_imports.size
               editor.remove_link_import_arg(link_node)
             end
           end
 
-          # TODO extract condition
-          next if editor.contains_link_usage?(link_node) || link_node.imports.size > removed_imports || dsl_parser.contains_object_usage?(link_node.name)
+          next if link_is_used?(link_node, remove_imports)
 
           editor.remove_link(link_node)
           removed_links += 1
@@ -50,6 +46,16 @@ module RubyLsp
         end
 
         editor.source
+      end
+
+      private
+
+      def import_is_used?(link_node, link_import)
+        editor.contains_link_import_usage?(link_node, link_import) || dsl_parser.contains_object_usage?(link_import)
+      end
+
+      def link_is_used?(link_node, remove_imports)
+        editor.contains_link_usage?(link_node) || link_node.imports.size > remove_imports.size || dsl_parser.contains_object_usage?(link_node.name)
       end
     end
   end