+++ ruby-lsp-ree/spec/completion/fn_completion_spec.rb
@@ -46,4 +46,62 @@ RSpec.describe "RubyLsp::Ree::CompletionListener" do
       expect(result.response.first.additional_text_edits.first.new_text).to match('link :seconds_ago')
     end
   end
+
+  it "returns params string based on signature" do
+    source =  <<~RUBY
+      class SomeClass
+        def something
+          func_with_three_args
+        end
+      end
+    RUBY
+
+    fn_source =  <<~RUBY
+      def func_with_three_args(string_field, id, third_arg)
+        puts id
+      end
+    RUBY
+
+    with_server(source) do |server, uri|
+      index_fn_from_source(server, fn_source)
+
+      send_completion_request(server, uri, { line: 2, character: 8 })
+      
+      result = server.pop_response
+
+      expect(result.response.size).to eq(1)
+      expect(result.response.first.label).to eq('func_with_three_args')
+      expect(result.response.first.text_edit.new_text).to eq('func_with_three_args(${1:string_field}, ${2:id}, ${3:third_arg})')
+    end
+  end
+
+  it "doesn't return params specification if arguments already entered" do
+    source =  <<~RUBY
+      class SomeClass
+        def something
+          func_with_three_args(some_obj.some_field, some_obj.id, third_arg)
+        end
+      end
+    RUBY
+
+    fn_source =  <<~RUBY
+      def func_with_three_args(string_field, id, third_arg)
+        puts id
+      end
+    RUBY
+
+    with_server(source) do |server, uri|
+      index_fn_from_source(server, fn_source)
+
+      send_completion_request(server, uri, { line: 2, character: 8 })
+      
+      result = server.pop_response
+
+      expect(result.response.size).to eq(1)
+      expect(result.response.first.label).to eq('func_with_three_args')
+      expect(result.response.first.text_edit.new_text).to eq('func_with_three_args')
+      expect(result.response.first.text_edit.range.start.character).to eq(4)
+      expect(result.response.first.text_edit.range.end.character).to eq(4 + 'func_with_three_args'.size)
+    end
+  end
 end